refactor: extract weather service module from registration wrapper
